requirements.txt
Flask==2.0.1
bcrypt==3.2.0
PyJWT==2.3.0
mysql-connector-python==8.0.27